requirements.txt
#requirement.txt
python-telegram-bot==20.6
requests
sqlite-utils
python-dotenv